Refactor block validators

+++ bitcoinkit/src/main/kotlin/io/horizontalsystems/bitcoinkit/bitcoincash/blocks/validators/EDAValidator.kt
@@ -5,18 +5,17 @@ import io.horizontalsystems.bitcoinkit.blocks.validators.BlockValidatorException
 import io.horizontalsystems.bitcoinkit.blocks.validators.IBlockValidator
 import io.horizontalsystems.bitcoinkit.crypto.CompactBits
 import io.horizontalsystems.bitcoinkit.models.Block
-import java.math.BigInteger
 
 // Emergency Difficulty Adjustment
-class EDAValidator(private val maxTargetBits: BigInteger, private val blockValidatorHelper: BitcoinCashBlockValidatorHelper) : IBlockValidator {
+class EDAValidator(private val maxTargetBits: Long, private val blockValidatorHelper: BitcoinCashBlockValidatorHelper) : IBlockValidator {
 
     override fun isBlockValidatable(block: Block, previousBlock: Block): Boolean {
         return true
     }
 
     override fun validate(block: Block, previousBlock: Block) {
-        if (previousBlock.bits.toBigInteger() == maxTargetBits) {
-            if (block.bits.toBigInteger() != maxTargetBits) {
+        if (previousBlock.bits == maxTargetBits) {
+            if (block.bits != maxTargetBits) {
                 throw BlockValidatorException.NotEqualBits()
             }
 
@@ -30,10 +29,10 @@ class EDAValidator(private val maxTargetBits: BigInteger, private val blockValid
         val mpt6blocks = blockValidatorHelper.medianTimePast(previousBlock) - blockValidatorHelper.medianTimePast(cursorBlock)
         if (mpt6blocks >= 12 * 3600) {
             val pow = CompactBits.decode(previousBlock.bits) shr 2
-            var powBits = CompactBits.encode(pow).toBigInteger()
+            var powBits = CompactBits.encode(pow)
             if (powBits > maxTargetBits)
                 powBits = maxTargetBits
-            if (powBits != block.bits.toBigInteger()) {
+            if (powBits != block.bits) {
                 throw BlockValidatorException.NotEqualBits()
             }
         } else if (previousBlock.bits != block.bits) {
